Update README.md
I added information about the deploy and status to the README menu. Now, it's possible to access the project link in the 'Deploy' section and check the deploy status in the 'Status' section. To do this, I added a level 3 header with the project link in the 'Deploy' section and included a Netlify badge with the deploy status in the 'Status' section.